bugfix. correctly handles filenames with spaces
Previously, if a file had a space in its filename and used cocoapods,
then objc-run would fail to build it correctly, as a result of trying to
copy to the unescaped filename, so that it copied only to the first
token of the filename.

This commit fixes that by wrapping the filename in quotes.