temporary directory is per script/project.
Previously, for scripts using Cocoapods, objc-run built an on-the-fly
project into a temporary project directory keyed only to the filename of
the script.

Now the temporary project directory is keyed to the script filename and
to exact version of objc-run itself.

Thus, objc-run will not use temporary assets that were created by
earlier versions of itself.